agent-matchmaker/requirements-app.txt
# Local Agent Matchmaker: pip install -r agent-matchmaker/requirements-app.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
